requirements.txt
tensorboardX==1.9      
tqdm==4.36.1 
pip==19.2.3 
opencv-python==4.2.0.32 
numpy==1.17.2
pandas==0.25.1
Pillow==6.2.0 
matplotlib==3.1.1
scikit-image==0.15.0       
easydict==1.8                
imageio==2.6.0
pyparsing==2.4.2             
python-dateutil==2.8.0                            
six==1.12.0                    
typing-extensions==3.7.4.1
scipy==1.3.1
torch==1.2.0
torchvision==0.4.0                    